Moved exec.js contents to tasks.js.
